+++ CHANGES.md
@@ -6,6 +6,7 @@ v1.2.1 (Month DD, YYYY)
 -----------------------
 
 - Updated number support to avoid locale issues (Issue #61)
+- Optimized string pool code.
 
 
 v1.2.0 (January 24, 2024)